luxury-brand/Final/luxury-brand/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Pillow==10.2.0
python-dotenv==1.0.0
Flask-CORS==4.0.0